Refactored @Game into the configurations package.
